refactor(config): 重构配置文件并添加新功能
- 重写 application.yml 文件结构，简化配置
- 新增开发、测试、生产环境的配置文件
- 添加文章控制器，实现文章管理相关功能

+++ ai-blog-springboot/src/main/java/com/ai/blog/entity/UserAuth.java
@@ -1,12 +1,12 @@
 package com.ai.blog.entity;
 
-import java.time.LocalDateTime;
-
 import com.baomidou.mybatisplus.annotation.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
 
 /**
  * 用户账号
@@ -26,7 +26,7 @@ public class UserAuth {
      * id
      */
     @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
+    private Long id;
 
     /**
      * 用户信息id
